// types.ts
export interface Attributes {
  corners: number; crossing: number; dribbling: number; finishing: number; firstTouch: number; freeKickTaking: number; heading: number; longShots: number; longThrows: number; marking: number; passing: number; penaltyTaking: number; tackling: number; technique: number;
  aggression: number; anticipation: number; bravery: number; composure: number; concentration: number; decisions: number; determination: number; flair: number; leadership: number; offTheBall: number; positioning: number; teamwork: number; vision: number; workRate: number;
  acceleration: number; agility: number; balance: number; jumpingReach: number; naturalFitness: number; pace: number; stamina: number; strength: number;
  aerialReach: number; commandOfArea: number; communication: number; eccentricity: number; handling: number; kicking: number; oneOnOnes: number; punching: number; reflexes: number; rushingOut: number; throwing: number;
  [key: string]: number;
}

export interface ManagerAttributes {
    attacking: number;
    defending: number;
    tactical: number;
    technical: number;
    mental: number;
    workingWithYoungsters: number;
    manManagement: number;
    determination: number;
    levelOfDiscipline: number;
    adaptability: number;
    [key: string]: number;
}

export interface AttributeSnapshot {
  id: string;
  date: string;
  attributes: Attributes;
  value?: number;
}

export interface ClubHistory {
  id: string;
  clubName: string;
  startYear: number;
  endYear: number | 'Present' | string;
  managedByUser: boolean;
  wasCaptain: boolean;
  transferFee?: number;
  isManagerTransfer?: boolean;
  isLoan?: boolean;
}

export interface SeasonStat {
  id: string;
  season: string;
  club: string;
  league?: string;
  apps: number;
  goals: number;
  assists: number;
  cleanSheets?: number;
  goalsConceded?: number;
  avgRating: number;
  honours: string[];
  pom: number;
}

export interface Player {
  id: string;
  firstName: string;
  lastName: string;
  knownAs: string;
  dateOfBirth: string;
  squadNumber: number;
  nationality: string;
  primaryPosition: string;
  currentClub: string;
  fmVersion: string;
  isRegen: boolean;
  isYouthIntake: boolean;
  isImported?: boolean;
  saveGameName: string;
  hasKnownAttributes: boolean;
  isFavourite: boolean;
  internationalCaps: number;
  internationalGoals: number;
  clubHistory: ClubHistory[];
  seasonStats: SeasonStat[];
  attributes: Attributes;
  attributeSnapshots: AttributeSnapshot[];
  originalManager: string;
  imageUrl?: string;
  facepackId?: string;
  currentValue?: number;
  customTags?: string[];
}
// ...
export interface Club {
  id: string;
  name: string;
  kitColors: {
    primary: string;
    secondary: string;
  };
  pattern: 'plain' | 'stripes' | 'hoops';
}
// ...
export interface CustomTag {
  id: string;
  name: string;
  color: string;
}